+++ backend/app/src/main/java/ch/zhaw/studyflow/Main.java
@@ -3,8 +3,8 @@ package ch.zhaw.studyflow;
 import ch.zhaw.studyflow.controllers.StudentController;
 import ch.zhaw.studyflow.domain.student.StudentManager;
 import ch.zhaw.studyflow.domain.student.impls.StudentManagerImpl;
-import ch.zhaw.studyflow.services.persistance.InMemorySettingsDao;
-import ch.zhaw.studyflow.services.persistance.InMemoryStudentDao;
+import ch.zhaw.studyflow.services.persistance.memory.InMemorySettingsDao;
+import ch.zhaw.studyflow.services.persistance.memory.InMemoryStudentDao;
 import ch.zhaw.studyflow.services.persistance.SettingsDao;
 import ch.zhaw.studyflow.services.persistance.StudentDao;
 import ch.zhaw.studyflow.webserver.WebServerBuilder;
