+++ backend/app/src/main/java/ch/zhaw/studyflow/domain/calendar/AppointmentManager.java
@@ -1,5 +1,7 @@
 package ch.zhaw.studyflow.domain.calendar;
 
+import ch.zhaw.studyflow.services.persistance.AppointmentDao;
+
 import java.util.Date;
 import java.util.List;
