+++ backend/app/src/main/java/ch/zhaw/studyflow/domain/calendar/CalendarManager.java
@@ -1,5 +1,7 @@
 package ch.zhaw.studyflow.domain.calendar;
 
+import ch.zhaw.studyflow.services.persistance.CalendarDao;
+
 public class CalendarManager {
     private final CalendarDao calendarDao;
