+++ backend/app/src/main/java/ch/zhaw/studyflow/services/persistance/sql/SQLAppointmentDao.java
@@ -1,4 +1,7 @@
-package ch.zhaw.studyflow.domain.calendar;
+package ch.zhaw.studyflow.services.persistance.sql;
+
+import ch.zhaw.studyflow.domain.calendar.Appointment;
+import ch.zhaw.studyflow.services.persistance.AppointmentDao;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
